PAs grand chose qui marche

+++ Moteurs/src/M_prop.c
@@ -20,7 +20,7 @@ void M_prop_init(void){
 	Stop();
 }
 void V_lent(){
-    SetDCPWM1(750);
+    SetDCPWM1((int)750);
 }
 void V_rapide(){
     SetDCPWM1(0x03FF);
